Interacciones medicamentosas implementado!!

+++ app/src/main/java/com/example/mitfg/ui/main/AlarmReceiver.kt
@@ -36,7 +36,7 @@ class AlarmReceiver : BroadcastReceiver() {
         val dosage = intent.getStringExtra("dosage")
 
         CoroutineScope(Dispatchers.IO).launch {
-            alarmRepository.addDosageToAlarm(idAlarm.toLong())
+            alarmRepository.addDosageToAlarm(idAlarm)
         }
 
         createSimpleNotification(context, idAlarm, medicineName, medicinePresentation, dosage)
